git-svn-id: svn://cherokee-project.com/cherokee/trunk@1693 5dc97367-97f1-0310-9951-d761b3857238

+++ cherokee/main_guardian.c
@@ -69,7 +69,7 @@ pid_file_clean (const char *pid_file)
 		return;
 	if (info.st_uid != getuid())
 		return;
-	if (info.st_size > sizeof("65535\r\n"))
+	if (info.st_size > (int) sizeof("65535\r\n"))
 		return;
 
 	unlink (pid_file);
